Stop adding FLAG_INSERT_AUTO_COPY for user-defined records
