+++ compiler/resolution/functionResolution.cpp
@@ -5774,38 +5774,6 @@ bool isInstantiation(Type* sub, Type* super) {
   return retval;
 }
 
-
-/************************************* | **************************************
-*                                                                             *
-*                                                                             *
-*                                                                             *
-************************************** | *************************************/
-
-bool requiresImplicitDestroy(CallExpr* call) {
-  bool retval = false;
-
-  if (FnSymbol* fn = call->resolvedFunction()) {
-    FnSymbol* parent = call->getFunction();
-
-    if (parent->hasFlag(FLAG_DONOR_FN)                        == false &&
-        isRecord(fn->retType)                                 == true  &&
-        fn->hasFlag(FLAG_NO_IMPLICIT_COPY)                    == false &&
-        fn->isIterator()                                      == false &&
-        fn->retType->symbol->hasFlag(FLAG_RUNTIME_TYPE_VALUE) == false &&
-        fn->hasFlag(FLAG_DONOR_FN)                            == false &&
-        fn->hasFlag(FLAG_INIT_COPY_FN)                        == false &&
-        fn->hasFlag(FLAG_AUTO_II)                             == false &&
-        fn->hasFlag(FLAG_CONSTRUCTOR)                         == false &&
-        fn->hasFlag(FLAG_TYPE_CONSTRUCTOR)                    == false &&
-        strcmp(fn->name, "=")                                 !=     0 &&
-        strcmp(fn->name, "_defaultOf")                        !=     0) {
-      retval = true;
-    }
-  }
-
-  return retval;
-}
-
 /************************************* | **************************************
 *                                                                             *
 *                                                                             *
